% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunEXTEND.r
\name{RunEXTEND}
\alias{RunEXTEND}
\title{EXTEND(EXpression based Telomerase ENzyme Detection)}
\usage{
RunEXTEND(data)
}
\arguments{
\item{infiles:}{Input = Gene expression matrix.Use RunEXTEND(data) command to execute the EXTEND function.}
}
\value{
Output would be generated in the form of a text file containing raw and normalized scores for Telomerase activity prediction.
}
\description{
This function takes the Gene expression data (in matrix format) as an input, where the Row Names correspond to Gene Names (as Gene Symbols) and the column Names correspond to sample Names.
}
\details{
RunEXTEND function processes the input matrix for calculation of Telomerase activity scores via internal modules.
}
